requirements.txt
cached_property
django
django-crontab
djangorestframework
psycopg2
requests
